+++ src/main/java/com/example/crudjdbc/controller/dto/ProdutoRequest.java
@@ -4,5 +4,5 @@ public record ProdutoRequest(
         String nome,
         Double preco,
         String descricao,
-        Long categoriaId) {
+        Integer categoriaId) {
 }
