+++ src/main/java/com/example/crudjdbc/repository/CategoriaRepository.java
@@ -3,5 +3,5 @@ package com.example.crudjdbc.repository;
 import com.example.crudjdbc.model.Categoria;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoriaRepository extends CrudRepository<Categoria, Long> {
+public interface CategoriaRepository extends CrudRepository<Categoria, Integer> {
 }
